move normalizeFilters to utils
