+++ apps/studio/src/common/utils.ts
@@ -151,7 +151,24 @@ export function safeSqlFormat(
   }
 }
 
+/** Join filters by AND or OR */
 export function joinFilters(filters: string[], ops: TableFilter[] = []): string {
   if (filters.length === 0) return ''
   return filters.reduce((a, b, idx) => `${a} ${ops[idx]?.op || 'AND'} ${b}`)
 }
+
+/** Get rid of invalid filters and parse if needed */
+export function normalizeFilters(filters: TableFilter[]) {
+  let normalized: TableFilter[] = [];
+  for (const filter of filters as TableFilter[]) {
+    if (!(filter.type && filter.field && filter.value)) continue;
+    if (filter.type === "in") {
+      const value = (filter.value as string).split(/\s*,\s*/);
+      normalized.push({ ...filter, value });
+    } else {
+      normalized.push(filter);
+    }
+  }
+  return normalized;
+}
+
